update
General information about a stock is stored in the stock_tickers table.

When a ticker (identified as $XXXX (dollar sign followed by an arbitrary number of letters), we check if that symbol in in the tickers table yet -if it's not there, we retrieve it from an external service. We then store the ticker ID (along with content_id and user_id) in the mentions table.

To come: Pull down current stock price when a ticker is mentioned.