Begin constructing modals
Template has a div in it called "modal-stub" - we use this to figure out where to append the Modals HTML in the page.

There is a new Routes file called Modals which contains not only the routes to pull out the modals, (which are rendered as HTML using parameters from a GET request), but also the routes to their respones.

So, modals.php contains routes for:

"/modals/report" to report questionable content
AND
"/report", which is for submitting the complaint.

The ReportContent modal is triggered by an onclick event that calls showReportContentModal()