Add heartbeat polling
Heartbeat middleware generates an ID for each pageload, and heartbeat.js pings the server every 10 seconds to update how long the page has been on screen.

Eventually, the server will be able to provide a more intelligent response to the client, indicating whether there are any new actions for the user to make.